Created Policy Section component
