+++ src/sections/Policy/index.ts
@@ -1,0 +1,2 @@
+import Policy from './Policy';
+export default Policy;
